reorganise fonctions
